amdended RepairClub code

+++ Workshop/TMA03SolJM/Computer.java
@@ -47,9 +47,9 @@ public class Computer extends Machine {
      * @return last service date of robot in unix time
      */
     @Override
-    public int lastServiced()
+    public long lastServiced()
     {
-        return Math.toIntExact(this.lastServiceDate);
+        return this.lastServiceDate;
     }
